% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{abs2rel}
\alias{abs2rel}
\title{abs2rel}
\usage{
abs2rel(gr, purity = NA, ploidy = NA, gamma = NA, beta = NA,
  field = "cn", field.ncn = "ncn", total = 1)
}
\arguments{
\item{gr}{GRanges input with meta data field corresponding to mean relative copy "mean" in that interval}

\item{purity}{purity of sample}

\item{ploidy}{ploidy of sample}

\item{gamma}{gamma fit of solution (over-rides purity and ploidy)}

\item{beta}{beta fit of solution (over-rides purity and ploidy)}

\item{field}{meta character specifying meta data field in "gr" variable from which to extract signal, default "mean"}

\item{field.ncn}{character specifying meta data field in "gr" variable from which to extract germline integer copy number, default "ncn", if doesn't exist, germline copy number is assumed to be zero}
}
\value{
numeric vector of integer copy numbers
}
\description{
rescales CN values from relative to "absolute" (i.e. per cancer cell copy) scale given purity and ploidy
By default, output is normalized to 1 (i.e. assumes that the total relative copy number signal mass over the genome is 1)
}
\details{
takes in gr with signal field "field"
}

