% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{adj2inc}
\alias{adj2inc}
\title{adj2inc}
\usage{
adj2inc(A)
}
\description{
converts adjacency matrix (of directed graph) into incidence matrix - ie
an nodes x edges matrix, for each edge i connecting node j to k, column i will have -1 at position
j and +1 at position k
}
