% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{all.paths}
\alias{all.paths}
\title{all.paths}
\usage{
\method{all}{paths}(A, all = F, ALL = F, sources = c(), sinks = c(),
  source.vertices = sources, sink.vertices = sinks, exclude = NULL,
  verbose = FALSE, ...)
}
\arguments{
\item{A}{nxn adjacency matrix}

\item{all}{logical flag, if all = T, will include all sources (parentless vertices) and sinks (childless vertices) in path computati}

\item{ALL}{logical flag, if ALL = T, will also include vertices without outgoing and incoming edges in paths}

\item{sources}{graph indices to treat as sources (by default is empty)}

\item{sinks}{graph indices to treat as sinks (by default is empty)}

\item{verbose}{logical flag}
}
\value{
list of integer vectors corresponding to indices in A (i.e. vertices)
$paths = paths indices
$cycles = cycle indices
}
\description{
Low level function to enumerate all elementary paths and cycles through graph
}
\details{
takes directed graph represented by n x n binary adjacency matrix  A and outputs all cycles and paths between source.vertices, sink.vertices
}

