% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{annotate.walks}
\alias{annotate.walks}
\title{annotate.walks}
\usage{
annotate.walks(walks, cds, promoters = NULL, filter.splice = T,
  verbose = F, prom.window = 1000, max.chunk = 1e+09, mc.cores = 1,
  exhaustive = FALSE)
}
\arguments{
\item{walks}{GRangesList of walks to query (eg from traversal of JaBbA object)}

\item{cds}{GRangesList of CDS annotation, one per transcript (each a coding region of an exon)}

\item{promoters}{GRanges of promoters (same length as transcript)}

\item{filter.splice}{flag whether to filter out splice variants of a given gene}

\item{verbose}{flag}

\item{prom.window}{window to use around each transcript to identify putative promoter if promoter is NULL}

\item{mc.cores}{number of cores to use}
}
\value{
a grl of putative fusions with annotations in values field:
$label
$type  e.g. promoter-fusion, 5-UTR fusion, In-frame fusion, 3' truncated fusion, 5' truncated fusion, in-frame poly-fusion, 3' truncated poly-fusion,
            5' truncated poly-fusion
$genes genes involved in fusion
$transcripts transcripts involved in fusion

annotates every possible altered transcript in region including
  - transcripts with truncated
}
\description{
Low level function to annotate walks (GRanges list) with cds / promoter annotations
}
\details{
given:
walks: input grl of walks on the genome
tx:  gr annotating transcript boundaries
or grl annotating exon level transcripts e.g. refgene or grl with 
grl-level meta data fields $s1, $s2, $e1, $e2 annotating start and end positions of transcipt and cds
respectively, $gene_sym representing gene label, $chr - chromosome, $str strand
and gr level features (exon_frame) annotating the frame (0,1,2) of the first exon position in a + transcript
and last exon position in a - transcript.
Assumes that exons are ordered in each grl item in the order of transcription (i.e. right most exon for negative strand transcripts)
(e.g. output of read_refGene(grl = T))
}

