% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch.n}
\alias{catch.n}
\title{catch.n}
\format{
data.frame with 1271 obs. of 7 variables
\describe{
\item{age}{age}
\item{year}{year}
\item{unit}{unit}
\item{season}{season}
\item{area}{area}
\item{iter}{iter}
\item{data}{data}
}
}
\usage{
catch.n
}
\description{
Ploy.sim catch by n
}
\keyword{datasets}
