% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{chromhmmpx_stub}
\alias{chromhmmpx_stub}
\title{run gene-chromhmm px analysis}
\usage{
chromhmmpx_stub(rg.span.file, chromhmm.file, this.ra.file, out.file,
  verbose = T, max.dist = 1e+06)
}
\description{
run gene-chromhmm px analysis
}

