% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{chromoplexy}
\alias{chromoplexy}
\title{Determines chromoplexy paths from standard JaBbA outpu}
\usage{
chromoplexy(kag = NULL, jab = NULL, sol = NULL, all = F, ref.only = F,
  filt.jab = T, paths = F, dist = 1000, cn.dist = dist, verbose = F,
  interval = 400, chunksize = 5000)
}
\arguments{
\item{jab}{JaBbA object}

\item{all}{logical flag whteher to enumerate all possible cycles, otherwise will return (an arbitrary) minimal decompositoin into the shortest "chains" of balanced rearrangements}

\item{paths}{logical flag if paths = T, will also try to compute paths (in addition to cycles), default = FALSE}

\item{dist}{maximum distance at which to cluster junctions, i.e. in which to consider a deletion or amplification bridge}

\item{cn.dist}{minimum distance at which to enforce copy concordance for deletion and amplification bridges}
}
\value{
paths and cycles of as list of vectors of aberrant edge index sequences,
aberrant edges refer to edges described in kag$ab.edges matrix
}
\description{
Outputs all chromoplexy paths and cycles
(i.e. paths and cycles in breakpoint graph) allowing quasi-reciprocal
rearrangements with amplification / deletion bridge distance threshold "dist"
}

