% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsy2jabba.R
\name{cmsy.bkprior}
\alias{cmsy.bkprior}
\title{cmsy.bkprior()}
\usage{
cmsy.bkprior(catch, bw = 3, prior.r = c(0.05, 0.5))
}
\arguments{
\item{catch}{catch time series, requires data.frame(year, catch)}

\item{bw}{kernel bandwith for catch smoothing (default=3)}

\item{prior.r}{prior.r TODO DOCUMENT}
}
\value{
list with initial start bk range and c(mu,cv)
}
\description{
approximates initial b/k prior using CMSY++15c.R
}
