% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsy2jabba.R
\name{cmsy.rprior}
\alias{cmsy.rprior}
\title{cmsy.prior()}
\usage{
cmsy.rprior(res)
}
\arguments{
\item{res}{Resiliance: Very low, Low, Medium, High}
}
\value{
r prior range
}
\description{
Computes r range from FishBase resiliance category
}
