% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{collapse.paths}
\alias{collapse.paths}
\title{collapse.paths}
\usage{
collapse.paths(G, verbose = T)
}
\description{
collapse simple paths in a graph G (adjacency matrix or igraph object)
returns m x m new adjacency matrix and map of old vertex id's to new ones
$adj = m x m matrix
#map = length n with indices 1 .. m
}
