% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{convex.basis}
\alias{convex.basis}
\title{karyoMIP utility functions}
\usage{
convex.basis(A, interval = 80, chunksize = 100, exclude.basis = NULL,
  exclude.range = NULL, maxchunks = Inf, verbose = F)
}
\details{
convex.basis

Outputs a matrix K of the convex basis of matrix A

i.e. each column x = K[,i] is a minimal solution (with respect to sparsity) to 
Ax = 0, x>=0

exclude.basis =  0, 1 matrix of dimension k x ncol(A) specifying k sparsity patterns that we would
like to exclude from the convex.basis.  This can speed up computation since any non-negative 
combinations of vectors that satisfy an exclusion property will also be excludable, and thus
we can remove such vectors as soon as we detect them..

exclude.range = 9, 1 matrix of dimension k x nrow(A) specifying k sparsity patterns that we would like
exclude, but these are specified in terms of the range of abs(A) .. i.e. we want to exclude all
basis vectors v such that nz(exclude.ranges[i, ]) C  nz(abs(A)*v)) for some pattern i.  Again
any non-neg linear comb of any intermediate-basis vector that satisfies this property will satisfy it,
as a result we can exclude these vectors when we see them.
}

