% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{cumseg}
\alias{cumseg}
\title{segment}
\usage{
cumseg(gr, field = "signal", log = T, type = "bic", alg = "stepwise",
  S = 1, verbose = F, mc.cores = 1, ...)
}
\description{
Wrapper around cumSeg to segment numeric data in an input GRanges with signal meta data field (e.g. $signal)
Returns a GRAnges of piecewise constant regions with their associated value
}

