% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_jabba.R
\name{fit_jabba}
\alias{fit_jabba}
\title{Fit JABBA model}
\usage{
fit_jabba(
  jbinput,
  ni = 30000,
  nt = 5,
  nb = 5000,
  nc = 2,
  init.values = FALSE,
  init.K = NULL,
  init.r = NULL,
  init.q = NULL,
  peels = NULL,
  save.all = FALSE,
  save.jabba = FALSE,
  save.csvs = FALSE,
  save.prjkobe = FALSE,
  output.dir = getwd(),
  quickmcmc = FALSE
)
}
\arguments{
\item{jbinput}{List of input variables as output by build_jabba()
MCMC settings}

\item{ni}{number of iterations}

\item{nt}{thinning interval of saved iterations}

\item{nb}{burn-in}

\item{nc}{number of mcmc chains
Initial values}

\item{init.values}{= FALSE,}

\item{init.K}{= NULL,}

\item{init.r}{= NULL,}

\item{init.q}{= NULL,# vector}

\item{peels}{= NULL, # retro peel option}

\item{save.all}{= FALSE}

\item{save.jabba}{= FALSE}

\item{output.dir}{path to save plot. default is getwd()}
}
\value{
A result list containing estimates of model input, settings and results
}
\description{
Fits JABBA model in JAGS and produce output object as list()
}
\examples{
data(iccat)
jbinput <- build_jabba(catch=iccat$bet$catch,cpue=iccat$bet$cpue,se=iccat$bet$se,model.type="Fox")
bet1 = fit_jabba(jbinput,quickmcmc=TRUE)
jbplot_cpuefits(bet1)
jbplot_ppdist(bet1)
par(mfrow=c(3,2),mar = c(3.5, 3.5, 0.5, 0.1))
jbplot_trj(bet1,type="B",add=T)
jbplot_trj(bet1,type="F",add=T)
jbplot_trj(bet1,type="BBmsy",add=T)
jbplot_trj(bet1,type="FFmsy",add=T)
jbplot_spphase(bet1,add=T)
jbplot_kobe(bet1,add=T)
}
