% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{fusions}
\alias{fusions}
\title{fusions}
\usage{
fusions(junctions = NULL, jab = NULL, cds = NULL, promoters = NULL,
  query = NULL, prom.window = 1000, verbose = T, max.chunk = 1e+10,
  cb.interval = 10000, cb.chunksize = 10000, cb.maxchunks = 1e+10,
  exhaustive = FALSE, debug = NULL, mc.cores = 1)
}
\arguments{
\item{junctions}{GRangesList of junctions (each a length 2 GRanges)}

\item{jab}{JaBbA object (overrides junctions input)}

\item{cds}{CDS annotations (GrangesList of transcript composed of coordinates coding regions of exons)}

\item{promoters}{GRanges of promoters (same length as transcript)}

\item{query}{optional query limiting walks to specific regions of interest}

\item{prom.window}{window to use around each transcript to identify putative promoter if promoter is NULL}
}
\value{
GRangesList of walks corresponding to transcript boundaires
}
\description{
annotates all gene fusions given an n x n adjacency matrix A of n genomic segments seg and grl of transcripts (eg output of read_RefGene)
seg must be (1) a tiling of the genome and (2) have copies of both + and - intervals for each genomic range (eg output of karyograph)
}
\details{
alternate input is a pile of junctions of ranges with strand orientation pointing AWAY from breakend

cds = gencode cds GRanges gff3 / gtf input

"gene_name" GRangesList meta data field is used in annotation and in not creating "splice" fusions that arise from different transcripts of the same gene.
}

