% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_jabba.R
\name{fw_jabba}
\alias{fw_jabba}
\title{fw_jabba()
External forward projections in JABBA}
\usage{
fw_jabba(
  jabba,
  quant = c("Catch", "F")[2],
  type = c("ratio", "msy", "abs")[2],
  initial = NULL,
  imp.yr = NULL,
  imp.values = seq(0.8, 1.2, 0.1),
  nyears = 10,
  nsq = 3,
  stochastic = c(TRUE, FALSE)[1],
  AR1 = c(TRUE, FALSE)[2],
  ndevs = 1,
  sigma.proc = NULL,
  rho = NULL,
  run = NULL,
  thin = 1
)
}
\arguments{
\item{jabba}{objects from fit_jabba() or list of fit_jabba() objects}

\item{quant}{quantity to forecast  c("Catch","F")
\itemize{
  \item Catch  
  \item F 
}}

\item{type}{choose the type of values provided  c("ratio","msy","abs")[1]
\itemize{
  \item ratio (relative to status quo Catch or F)   
  \item msy (relative to Fmsy or MSY)
  \item abs (input values are taken as absolute values)
}}

\item{initial}{value or vector Catch or F values, default takes mean over recent 3 yrs}

\item{imp.yr}{management implementation year in number of years after last data year}

\item{imp.values}{vector Catch or F scenarios provide as absolute or ratios}

\item{nyears}{number of forecast years}

\item{stochastic}{if FALSE, process error sigma.proc is set to zero}

\item{AR1}{if TRUE, projection account auto correlation in the process devs}

\item{ndevs}{number years on the tail to set initial proc.error for forecasting}

\item{sigma.proc}{option to specify the process error other than the posterior estimate}

\item{rho}{if AR1 = TRUE, the autocorrelation coefficient is estimated from the proc devs}

\item{run}{option to assign a scenario name other than specified in build_jabba()}

\item{thin}{option to thin the posterior at rates > 1}
}
\value{
data.frame of kobe posterior model + forecast scenarios
}
\description{
fw_jabba()
External forward projections in JABBA
}
