% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_jabba.R
\name{fw_jabba_ctrl}
\alias{fw_jabba_ctrl}
\title{fw_jabba_ctrl()
External forward projections in JABBA}
\usage{
fw_jabba_ctrl(
  jabba,
  fwdctrl = NULL,
  quant = c("Catch", "F", "Fmsy", "Fsq", "Csq")[3],
  initial = NULL,
  imp.yr = NULL,
  imp.values = seq(0.8, 1.2, 0.1),
  nyears = 5,
  nsq = 3,
  stochastic = c(TRUE, FALSE)[1],
  AR1 = c(TRUE, FALSE)[2],
  ndevs = 1,
  sigma.proc = NULL,
  rho = NULL,
  run = NULL,
  seed = 123,
  thin = 1
)
}
\arguments{
\item{jabba}{objects from fit_jabba() or list of fit_jabba() objects}

\item{fwdctrl}{control table - overwrites quants, years, values}

\item{quant}{quantity to forecast  c("Catch","F")
\itemize{
  \item Catch (absolute Catch)  
  \item F (absolute F)
  \item Fmsy (relative to Fmsy)
  \item Fsq (relative to status quo F)   
  \item Csq (relative to status quo Catch) 
}}

\item{initial}{value or vector Catch or F values, default takes mean over recent 3 yrs}

\item{imp.yr}{management implementation year in number of years after last data year}

\item{imp.values}{vector Catch or F scenarios provide as absolute or ratios}

\item{nyears}{number of forecast years}

\item{stochastic}{if FALSE, process error sigma.proc is set to zero}

\item{AR1}{if TRUE, projection account auto correlation in the process devs}

\item{ndevs}{number years on the tail to set initial proc.error for forecasting}

\item{sigma.proc}{option to specify the process error other than the posterior estimate}

\item{rho}{if AR1 = TRUE, the autocorrelation coefficient is estimated from the proc devs}

\item{run}{option to assign a scenario name other than specified in build_jabba()}

\item{seed}{default 123}

\item{thin}{option to thin the posterior at rates > 1}
}
\value{
data.frame of kobe posterior model + forecast scenarios
}
\description{
fw_jabba_ctrl()
External forward projections in JABBA
}
