% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_utils.R
\name{get_beta}
\alias{get_beta}
\title{gets beta prior paramenters}
\usage{
get_beta(mu, CV, Min = 0, Prior = "x", Plot = FALSE)
}
\arguments{
\item{mu}{mean of beta distribution}

\item{CV}{cv of beta distribution}

\item{Min}{min of x-axis range (default = 0)}

\item{Prior}{X-axis label}

\item{Plot}{c(TRUE,FALSE)}
}
\value{
a and b parameter (shape and scale)
}
\description{
gets beta prior paramenters and plots distribution
}
