% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_utils.R
\name{get_gamma}
\alias{get_gamma}
\title{gets gamma prior paramenters}
\usage{
get_gamma(mu, CV, Prior = "x", Plot = FALSE)
}
\arguments{
\item{mu}{mean of gamma distribution}

\item{CV}{cv of gamma distribution}

\item{Prior}{X-axis label}

\item{Plot}{c(TRUE,FALSE)}
}
\value{
a and b parameter (shape and scale)
}
\description{
gets gamma prior paramenters and plots distribution
}
