% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast_jabba.R
\name{hindcast_jabba}
\alias{hindcast_jabba}
\title{hindcast_jabba()}
\usage{
hindcast_jabba(
  jbinput,
  fit,
  ni = NULL,
  nt = NULL,
  nb = NULL,
  nc = NULL,
  quickmcmc = TRUE,
  init.values = TRUE,
  peels = 1:5,
  verbose = FALSE
)
}
\arguments{
\item{jbinput}{object from build_jabba()}

\item{ni}{number of iterations}

\item{nt}{thinning interval of saved iterations}

\item{nb}{burn-in}

\item{nc}{number of mcmc chains}

\item{quickmcmc}{Reduces MCMC iters for hindcasting reference run
Initial values}

\item{init.values}{if TRUE init values from fit are used}

\item{peels}{sequence of retrospective peels default 1:5}

\item{verbose}{if FALSE run silent}

\item{jbfit}{fitted model from fit_jabba
MCMC settings}
}
\value{
hc containing estimates of key joint results from all hindcast run
}
\description{
Wrapper to fit retrospectives
}
\examples{
data(iccat)
whm = iccat$whm
# ICCAT white marlin setup
jb = build_jabba(catch=whm$catch,cpue=whm$cpue,se=whm$se,assessment="WHM",scenario = "BaseCase",model.type = "Pella",r.prior = c(0.181,0.18),BmsyK = 0.39,igamma = c(0.001,0.001))
fit = fit_jabba(jb,quickmcmc=TRUE,verbose=TRUE)
hc = hindcast_jabba(jbinput=jb,fit=fit,peels=1:5)
jbplot_retro(hc)
jbplot_hcxval(hc,index=c(8,11))
hc.ar1 = jbhcxval(hc,AR1=TRUE) # do hindcasting with AR1
jbplot_hcxval(hc.ar1,index=c(8,11))
}
