% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{index}
\alias{index}
\title{index}
\format{
data.frame of 46 obs. of 8 variables
\describe{
\item{age}{age}
\item{year}{year}
\item{unit}{unit}
\item{season}{season}
\item{area}{area}
\item{iter}{iter}
\item{data}{data}
\item{qname}{qname}
}
}
\usage{
index
}
\description{
Poly.sim index
}
\keyword{datasets}
