% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{jabba.alleles}
\alias{jabba.alleles}
\title{jabba.alleles}
\usage{
jabba.alleles(jab, het.sites, alt.count.field = "alt.count.t",
  ref.count.field = "ref.count.t", baf.field = "baf.t", split.ab = F,
  uncoupled = FALSE, conservative = FALSE, verbose = F)
}
\arguments{
\item{jab}{JaBbA object}

\item{het.sites}{GRanges with meta data fields (see below) for alt and rref count}

\item{alt.count.field}{character specifying alt.count meta data field in input het.sites (default alt.count.t)}

\item{ref.count.field}{character specifying ref.count meta data field in input het.sites (default ref.count.t)}

\item{split.ab}{logical flag whether to split aberrant segmetns (segmentss with ab edge entering or leaving prior to computing allelic states (default FALSE)}

\item{uncoupled}{logical flag whether to not collapse segments after inferring MLE estimate (default FALSE), if FALSE will try to merge adjacent segments and populate allele-specific junctions with copy numbers on the basis of the MLE fit on individual allelic segments}

\item{conservative}{if TRUE then will leave certain allelic segments "unphased" if one cannot sync the high / low interval state with the incoming and / or outgoing junction state}

\item{verbose}{logical flag}
}
\value{
list with following fields:
$segstats = GRanges of input segments with $cn.high and $cn.low segments populated
$asegstats = GRanges of allelic segments (length is 2*length(segstats)) with high and low segments each having $cn, this is a "melted" segstats GRAnges
$atd = gTrack of allelic segments and supporting input het.sites
$aadj = allelic adjacency matrix of allele specific junctions
$ab.ix = indices of aberrant edges in $aadj
$ref.ix = indices of reference edges in $aadj
}
\description{
Populates allelic value s for JaBbA object.  This does not explicitly impose junction balance constraints on alleles, but rather just computes
the maximum likelihood estimate given allelic counts and the inferred total copy number on a given segment according to JaBbA
}
