% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{JaBbA.digest}
\alias{JaBbA.digest}
\title{JaBbA.digest}
\usage{
JaBbA.digest(jab, kag = NULL, verbose = T, keep.all = T)
}
\arguments{
\item{jab}{JaBbA object "undigested"}

\item{kag}{karyograph (original karyograph input to JaBbA), if NULL then will "redigest" JaBbA object}

\item{verbose}{logical flag}

\item{keep.all}{keep.all (default TRUE) whether to keep 0 copy junctions or collapse segments across these as well}
}
\description{
JaBbA.digest
}
\details{
processes JaBbA object
(1) collapsing segments with same copy number that lack loose ends
(2) (optional) +/- adds segments correponding to loose ends
(3) outputting edges data frame with colors, and other formatting information
(4) outputting junctions GRangesList with copy number, color, lty and other plotting components

TODO: replace with proper object instantiator
}
