% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{jabba.dist}
\alias{jabba.dist}
\title{jabba.dist}
\usage{
jabba.dist(jab, gr1, gr2, matrix = T, max.dist = Inf,
  include.internal = TRUE, EPS = 1e-09)
}
\arguments{
\item{jab}{JaBbA object}

\item{gr1}{interval set 1 GRanges}

\item{gr2}{interval set 2 GRanges}

\item{matrix}{flag whteher to output a matrix}

\item{max.dist}{numeric (default = Inf), if non-infinity then output will be a sparse matrix with all entries that are greater than max.dist set to zero}
}
\value{
a length(gr1) x length(gr2) matrix whose entries ij store the distance between
the 3' end of gr1[i] and 5' end of gr2[j]
}
\description{
Computes distance between pairs of intervals on JaBbA graph
}
\details{
Given "jabba" object and input granges gr1 and gr2 of (signed) intervals
}

