% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{jabba.hood}
\alias{jabba.hood}
\title{jabba.hood}
\usage{
jabba.hood(jab, win, d = 0, k = NULL, pad = 0, ignore.strand = TRUE,
  bagel = FALSE, verbose = FALSE)
}
\arguments{
\item{jab}{JaBbA object}

\item{win}{GRanges of window of interest}

\item{d}{= distance in coordinates on graph}

\item{k}{Neighborhood on graph around window of interest to query}

\item{pad}{pad level at which to collapse nearly reference adjacent intervals}
}
\value{
a reduced set of windows within neighborhood k
of seed on the graph (only includes edges with weight !=0)
}
\description{
Given JaBbA  object
and seed window "win", outputs a reduced set of windows within neighborhoof of n coordinate (ork nodes)
within the seed region(s) on the graph (only includes edges with weight !=0)
}
