% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{jabba2vcf}
\alias{jabba2vcf}
\title{jabba2vcf}
\usage{
jabba2vcf(jab, fn = NULL, sampleid = "sample", hg = skidb::read_hg(fft =
  T), cnv = FALSE)
}
\arguments{
\item{jab}{JaBbA object}

\item{fn}{output file name}

\item{sampleid}{sample id}

\item{hg}{human genome as BSgenome or ffTrack}

\item{cnv}{flag whether to dump in CNV format}
}
\value{
returns character string or writes to file if specified
}
\description{
Converts jabba output to vcf file according to 4.2 "BND" syntax
}

