% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_jabba.R
\name{jabba_fw}
\alias{jabba_fw}
\title{jabba_fw()
External forward projections in JABBA}
\usage{
jabba_fw(
  jabba,
  nyears = 10,
  imp.yr = 2,
  status.quo = NULL,
  values = seq(0.6, 1.2, 0.1),
  quant = c("Catch", "F")[1],
  type = c("ratio", "msy", "abs")[1],
  stochastic = c(TRUE, FALSE)[1],
  AR1 = c(TRUE, FALSE)[2],
  sigma.proc = NULL,
  rho = NULL,
  run = NULL,
  thin = 1
)
}
\arguments{
\item{jabba}{objects from fit_jabba() or list of fit_jabba() objects}

\item{values}{vector Catch or F values provide as absolute or ratios}

\item{quant}{quantity to forecast  c("Catch","F")
\itemize{
  \item Catch  
  \item F 
}}

\item{type}{choose the type of values provided  c("ratio","msy","abs")[1]
\itemize{
  \item ratio (relative to status quo Catch or F)   
  \item msy (relative to Fmsy or MSY)
  \item abs (input values are taken as absolute values)
}}

\item{stochastic}{if FALSE, process error sigma.proc is set to zero}

\item{AR1}{if TRUE, projection account auto correlation in the process devs}

\item{sigma.proc}{option to specify the process error other than the posterior estimate}

\item{rho}{if AR1 = TRUE, the autocorrelation coefficient is estimated from the proc devs}

\item{run}{option to assign a scenario name other than specified in build_jabba()}

\item{thin}{option to thin the posterior at rates > 1}
}
\value{
data.frame of kobe posterior model + forecast scenarios
}
\description{
jabba_fw()
External forward projections in JABBA
}
