% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jabba_plots}
\alias{jabba_plots}
\title{wrapper jbplot function}
\usage{
jabba_plots(jabba, output.dir = getwd(), as.png = TRUE, statusplot = "kobe")
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}
}
\description{
plots all routine JABBA plot to output.dir if as.png=TRUE (default)
}
