% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_utils.R
\name{jb_hockeystick}
\alias{jb_hockeystick}
\title{jb_hockeystick()}
\usage{
jb_hockeystick(
  fwd,
  fyr = 3,
  btrigger = 0.5,
  fmsy = 1,
  bmin = 0,
  fmin = 0.001,
  uncertainty = F
)
}
\arguments{
\item{fwd}{output jabba_forward_ctrl()}

\item{fyr}{forecast year for evaluating B/Bmsy (e.g. 1,2,3)}

\item{btrigger}{b as ratio of B/Bmsy, if b below btrigger F is reduced}

\item{fmsy}{target ratio of F/Fmsy (default =1)}

\item{bmin}{Option for de facto fishing closure (e.g. bmin = blim) description}

\item{fmin}{F/Fmsy below bim}
}
\value{
ratio F/Fmsy
}
\description{
F-based Hockey-Stick Harvest Control rule
}
