% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{jbaMIP.allelic}
\alias{jbaMIP.allelic}
\title{jbaMIP.allelic}
\usage{
jbaMIP.allelic(adj, segstats, purity, gamma, partition = T,
  slack.prior = 0.001)
}
\arguments{
\item{adj}{adjacency matrix populated with total copy counts on junctions}

\item{segstats}{granges tiling genome populated with total copy counts on interval, mu_high, sd_high, mu_low, sd_low variables on alleles}

\item{purity}{purity from solution}

\item{gamma}{gamma param from jbaMIP}

\item{slack.prior}{1/slack.prior = penalty for each slack i.e. loose end copy in solution}
}
\value{
list with fields
$segstats = input segstats annotated with fitted cn.high, cn.low columns
$asegstats = output "allelic" segstats, with $cn, $parent.node, $eslack.in, $eslack.out, $phased fields filled in
$adj = output length(segstats) x length(segstats) x 2  "allelic" adjacency matrix with inferred allelic copy numbers on edges
$aadj = flattened output 2*length(segstats) x 2* length(segstats)  "allelic" adjacency matrix with inferred allelic copy numbers on edges
}
\description{
Takes adj and segstats from output from jbaMIP and
a granges of het.sites with $ref.count and $alt.count
}
\details{
assumes segstats has fields $cn populated
and adj has copy states
}
