% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{jbaMIP.process}
\alias{jbaMIP.process}
\title{jbaMIP.process}
\usage{
jbaMIP.process(sol, allelic = F)
}
\arguments{
\item{sol}{JaBbA object}

\item{allelic}{logical flag specifying whether object is allelic}
}
\value{
list with items:
$B incidence matrix of augmented graph (including slack vertices) (vertices x edges)
rownames of $B are vertex names of $G and colnames of B are named with character version of their $G indices
(i.e. column order of B  respects the original edge order in the solution)

$e edge constraints for downstream karyoMIP, i.e the copy numbers at the edges
$e.ij numedges x 2 vertex pair matrix denoting what are the vertex pairs corresponding to the cols of $B and entries of $e, $eclass, $etype etc
$eclass id for each unique edge / anti-edge equivalence class
$etype specifies whether edge is slack or nonslack
}
\description{
process jbaMIP solution "sol" given original graph "g" (karyograph() list output)
into JaBbA object
}
\details{
output is
}
