% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbhcxval.R
\name{jbhcxval}
\alias{jbhcxval}
\title{jbhcxval()
additional hindcast options with external foreward projections}
\usage{
jbhcxval(
  hindcasts,
  stochastic = c(TRUE, FALSE)[1],
  AR1 = c(TRUE, FALSE)[1],
  sigma.proc = NULL,
  rho = NULL,
  ndevs = 1,
  run = NULL
)
}
\arguments{
\item{stochastic}{if FALSE, process error sigma.proc is set to zero}

\item{AR1}{if TRUE, projection account auto correlation in the process devs}

\item{sigma.proc}{option to specify the process error other than the posterior estimate}

\item{rho}{if AR1 = TRUE, the autocorrelation coefficient is estimated from the proc devs}

\item{ndevs}{number years on the tail to set initial proc.error for forecasting}

\item{run}{option to assign a scenario name other than specified in build_jabba()}

\item{hc}{object (list of models) from hindcast_jabba()}

\item{thin}{option to thin the posterior at rates > 1}
}
\value{
data.frame of kobe posterior model + forecast scenarios
}
\description{
jbhcxval()
additional hindcast options with external foreward projections
}
\examples{
data(iccat)
whm = iccat$whm
# ICCAT white marlin setup
jb = build_jabba(catch=whm$catch,cpue=whm$cpue,se=whm$se,assessment="WHM",scenario = "BaseCase",model.type = "Pella",r.prior = c(0.181,0.18),BmsyK = 0.39,igamma = c(0.001,0.001))
fit = fit_jabba(jb,quickmcmc=TRUE,verbose=TRUE)
hc = hindcast_jabba(jbinput=jb,fit=fit,peels=1:5)
jbplot_retro(hc)
jbplot_hcxval(hc,index=c(8,11))
hc.ar1 = jbhcxval(hc,AR1=TRUE) # do hindcasting with AR1
jbplot_hcxval(hc.ar1,index=c(8,11))
}
