% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R, R/selex.R
\name{jbio}
\alias{jbio}
\title{jbio()}
\usage{
jbio(
  amin = 0,
  amax = 10,
  nsexes = 2,
  grow = NULL,
  model = "vb",
  Linf = c(80, 65),
  k = c(0.2, 0.25),
  t0 = c(-0.5, -0.7),
  L0 = NULL,
  aW = 0.01,
  bW = 3.04,
  mat = c(40, 45, 0),
  M = c(0.3, 0.4),
  h = 0.7,
  fec = NA,
  empty = FALSE
)

jbio(
  amin = 0,
  amax = 10,
  nsexes = 2,
  grow = NULL,
  model = "vb",
  Linf = c(80, 65),
  k = c(0.2, 0.25),
  t0 = c(-0.5, -0.7),
  L0 = NULL,
  aW = 0.01,
  bW = 3.04,
  mat = c(40, 45, 0),
  M = c(0.3, 0.4),
  h = 0.7,
  fec = NA,
  empty = FALSE
)
}
\arguments{
\item{amin}{minimum age}

\item{amax}{maximum age}

\item{nsexes}{number of sexes}

\item{grow}{optional output list from \code{jgrow()}}

\item{model}{growth model c("vb","schnute") if \code{grow=NULL}}

\item{Linf}{asymptotic length}

\item{k}{growth coefficient}

\item{t0}{theoretical age at zero length}

\item{L0}{optional alternative to \code{t0}}

\item{aW}{length-weight parameter a}

\item{bW}{length-weight parameter b}

\item{mat}{maturity parameters c(m50,m95,type), where type = 0 for length and 1 for age}

\item{M}{natural mortality}

\item{h}{steepness}

\item{fec}{optional fecundity at age}

\item{empty}{logical option to create empty object}

\item{Loo}{asymptotic length}
}
\value{
jbio List of life history vectors and matrices

list with age, La, Wa, Mat, M, h and optional fec
}
\description{
Creates a list for JABBA-Select stock object as input for the ASEM

Creates a stock object for JABBA-Select life history
}
\examples{
stk <- jbio(
  amin = 0, amax = 20, nsexes = 2,
  Linf = c(80,65), k = c(0.2,0.25), t0 = c(-0.5,-0.7),
  aW = 0.01, bW = 3.04,
  mat = c(40,45,0), M = c(0.3,0.4), h = 0.7
)
head(stk$La)

g <- jgrow(list(age = 0:20, nsexes = 2,
                Linf = c(80,65), k = c(0.2,0.25), t0 = c(-0.5,-0.7)))
stk2 <- jbio(grow = g, aW = 0.01, bW = 3.04,
             mat = c(40,45,0), M = c(0.3,0.4), h = 0.7)
}
