% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{jbleslie}
\alias{jbleslie}
\title{jbleslie}
\usage{
jbleslie(bio)
}
\arguments{
\item{bio}{list of history parameters 
\itemize{
  \item if fecundity is provided (e.g. pups per females) it will be used directly  
  \item if fecundity is NULL, a BevHolt as a function is used to compute the annual reproductive rate
}}
}
\value{
list r, gt
}
\description{
Function to compute r, generation time (gt) based on female life history
}
\examples{
#Blond ray
# Use fecundity: 30 per female
bio.fec = jbio(amin = 0,amax = 20,Loo = c(154.7),k=c(0.129),t0=c(-1),aW = 0.01,bW=3.04,mat=c(5.5,6,1),M=c(0.26),fec=32,empty=FALSE)
jbleslie(bio.fec)
# Use steepness: h = 0.5
bio.h = jbio(amin = 0,amax = 20,Loo = c(154.7),k=c(0.129),t0=c(-1),aW = 0.01,bW=3.04,mat=c(5.5,6,1),M=c(0.26),h=0.5,empty=FALSE)
jbleslie(bio.h)
#
# Thornback ray
# Use fecundity: 100 per female
bio.fec = jbio(amin = 0,amax = 18,Loo = c(117.6),k=c(0.16),t0=c(-1),aW = 0.01,bW=3.04,mat=c(5.3,5.8,1),M=c(0.32),fec=100,empty=FALSE)
jbleslie(bio.fec)
# Use steepness: h = 0.5
bio.h = jbio(amin = 0,amax = 18,Loo = c(117.6),k=c(0.16),t0=c(-1),aW = 0.01,bW=3.04,mat=c(5.3,5.8,1),M=c(0.32),h=0.5,empty=FALSE)
jbleslie(bio.h)

}
