% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbpar}
\alias{jbpar}
\title{jbpar()}
\usage{
jbpar(
  mfrow = c(1, 1),
  mex = 0.8,
  plot.cex = 0.8,
  mai = c(0.5, 0.5, 0.15, 0.15),
  omi = c(0.2, 0.2, 0.2, 0),
  mar = c(2.5, 2.5, 0.7, 0.7),
  labs = TRUE
)
}
\arguments{
\item{mfrow}{determines plot frame set up}

\item{plot.cex}{cex graphic option}
}
\description{
Set the par() to options suitable for JARA multi plots
}
