% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_bprior}
\alias{jbplot_bprior}
\title{Plot of biomass depletion prior vs posterios}
\usage{
jbplot_bprior(
  jabba,
  output.dir = getwd(),
  as.png = FALSE,
  add = FALSE,
  width = 5,
  height = 3.5
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{add}{if true don't call par() to allow construction of multiplots}

\item{width}{plot width}

\item{height}{plot hight}
}
\description{
prior is specified as b/k or b/bmsy in build_jabba()
}
