% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbplot_ensemble.R
\name{jbplot_ensemble}
\alias{jbplot_ensemble}
\title{jbplot_ensemble()}
\usage{
jbplot_ensemble(
  kb,
  subplots = c("B", "H", "stock", "harvest", "Bdev", "Catch", "BB0")[1:6],
  joint = FALSE,
  quantiles = c(0.025, 0.975),
  kbout = FALSE,
  ylabs = NULL,
  plot = TRUE,
  print = FALSE,
  png = print,
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  tickEndYr = FALSE,
  xlim = NULL,
  ylimAdj = 1.05,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "l",
  uncertainty = TRUE,
  legend = TRUE,
  legendlabels = "default",
  legendloc = "topleft",
  legendorder = "default",
  legendncol = 1,
  legendcex = 0.7,
  legendsp = 0.8,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  shadecol = NULL,
  shadealpha = 0.3,
  new = TRUE,
  add = TRUE,
  run = NULL,
  fmax = 5
)
}
\arguments{
\item{kb}{objects from fit_jabba(),jabba_fw(), list of fit_jabba() or fit_jabba()$kbtrj}

\item{subplots}{option to c("stock","harvest","B","H","Bdev","Catch","BB0") 
\itemize{
  \item stock (B/Bmsy)  
  \item harvest (F/Fmsy) 
  \item B (Biomass) 
  \item H (Harvest rate)
  \item Bdev (Process Deviations)
  \item Catch
  \item BB0 (B/K) 
}}

\item{joint}{if true it creates a joint ensemble from list of fit_jabba()}

\item{kbout}{if TRUE, produces the kb data.frame as output}

\item{ylabs}{yaxis labels for quants
final year of values to show for each model. By default it is set to the}

\item{col}{Optional vector of colors to be used for lines. Input NULL}

\item{pch}{Optional vector of plot character values}

\item{lty}{Optional vector of line types}

\item{lwd}{Optional vector of line widths}

\item{tickEndYr}{TRUE/FALSE switch to turn on/off extra axis mark at final
year in timeseries plots.}

\item{xlim}{= NULL range of years}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space}

\item{xaxs}{Choice of xaxs parameter (see ?par for more info)}

\item{yaxs}{Choice of yaxs parameter (see ?par for more info)}

\item{type}{Type parameter passed to points (default 'o' overplots points on
top of lines)}

\item{legend}{Add a legend?}

\item{legendlabels}{Optional vector of labels to include in legend.}

\item{legendloc}{Location of legend. Either a string like "topleft" or a vector
of two numeric values representing the fraction of the maximum in the x and y
dimensions, respectively. See ?legend for more info on the string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have
the legend display the model names in an order that is different than that
which is represented in the summary input object.}

\item{legendncol}{Number of columns for the legend.}

\item{legendsp}{Space between legend labels}

\item{pwidth}{Width of plot}

\item{pheight}{Height of plot}

\item{punits}{Units for PNG file}

\item{res}{Resolution for PNG file}

\item{ptsize}{Point size for PNG file}

\item{cex.main}{Character expansion for plot titles}

\item{plotdir}{Directory where PNG or PDF files will be written. By default
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to PNG or PDF file names.
It will be separated from default name by an underscore.}

\item{par}{list of graphics parameter values passed to par() function}

\item{verbose}{Report progress to R GUI?}

\item{shadecol}{uncertainty shading of hcxval horizon}

\item{shadealpha}{Transparency adjustment used to make default shadecol}

\item{new}{Create new empty plot window}

\item{add}{surpresses par() to create multiplot figs}

\item{run}{name for single models or joint ensembles}

\item{legendcex=1}{Allows to adjust legend cex}
}
\description{
Plots plots JABBA ensemble models + projections - joint or by run
}
\author{
Mostly adopted from ss3diags::SSplotEnsemble
}
