% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_hcxval}
\alias{jbplot_hcxval}
\title{jbplot_hcxval}
\usage{
jbplot_hcxval(
  hc,
  output.dir = getwd(),
  as.png = FALSE,
  single.plots = FALSE,
  add = FALSE,
  width = NULL,
  height = NULL,
  minyr = NULL,
  cols = NULL,
  legend.loc = "topright",
  legend.cex = 0.8,
  legend.add = TRUE
)
}
\arguments{
\item{hc}{output ojejct from jabba_hindcast}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{add}{if TRUE plots par is only called for first plot}

\item{width}{plot width}

\item{height}{plot hight}

\item{minyr}{minimum year shown in plot}

\item{cols}{option to add colour palette}

\item{legend.loc}{location of legend}

\item{legend.cex}{size of legend}

\item{legend.add}{show legend}
}
\value{
hcxval statistics by index: MASE, MAE.PR predition residuals,MAE.base for random walk, n.eval obs evaluated
}
\description{
Plots and summarizes results from one step head hindcast cross-validation using the output form jabba_hindcast
}
