% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_indices}
\alias{jbplot_indices}
\title{jbplot_indices}
\usage{
jbplot_indices(
  jabbainput,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  plot.cex = 0.9,
  xlab = "Year",
  ylab = "Abundance index",
  add = FALSE,
  add.legend = TRUE,
  cols = NULL
)
}
\arguments{
\item{jabbainput}{output list from build_jabba()}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{plot.cex}{cex graphic option}

\item{xlab}{option to change x-axis label}

\item{ylab}{option to change y-axis label}

\item{add}{if TRUE par is not called to enable manual multiplots}

\item{add.legend}{option to add legend}

\item{cols}{option to choose own colour palette}
}
\description{
Plots indices with assumed mimumum observation errors
}
