% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_logfits}
\alias{jbplot_logfits}
\title{log(CPUE) fits}
\usage{
jbplot_logfits(
  jabba,
  index = NULL,
  output.dir = getwd(),
  add = FALSE,
  as.png = FALSE,
  single.plots = add,
  width = NULL,
  height = NULL,
  plotCIs = TRUE
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{index}{option to plot specific indices (numeric & in order)}

\item{output.dir}{directory to save plots}

\item{add}{if TRUE is surpresses par() within the plotting function}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{width}{plot width}

\item{height}{plot hight}

\item{plotCIs}{plots error bars on CPUE observations}
}
\description{
Plot of fitted CPUE indices on log-scale (r4ss-style)
}
