% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_mcmc}
\alias{jbplot_mcmc}
\title{Plot mcmc chains}
\usage{
jbplot_mcmc(
  jabba,
  output.dir = getwd(),
  as.png = FALSE,
  mfrow = c(round((ncol(jabba$pars_posterior))/3 + 0.33, 0), 3),
  width = 8,
  height = 2.5 * round(jabba$pars_posterior/3, 0)
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{mfrow}{set up plot frame}

\item{width}{plot width}

\item{height}{plot hight}

\item{add}{if true don't call par() to allow construction of multiplots}
}
\description{
MCMC chains of esimated parameters: K, r, m (shape), psi (depletion) and variances
}
