% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_PPC}
\alias{jbplot_PPC}
\title{jrplot_PPC}
\usage{
jbplot_PPC(
  jabba,
  joint.ppc = FALSE,
  thin.plot = TRUE,
  output.dir = getwd(),
  as.png = FALSE,
  single.plots = add,
  width = NULL,
  height = NULL,
  ylab = expression(paste("Predicted D(", chi^2, ")")),
  xlab = expression(paste("Realized D(", chi^2, ")")),
  plot.cex = 0.8,
  index = NULL,
  index.label = TRUE,
  add = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{joint.ppc}{FALSE/TRUE, if TRUE a joint Posterior Predictive Check (indices combined) is plotted}

\item{thin.plot}{TRUE/FALSE thinning option for plotting}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{width}{plot width}

\item{height}{plot hight}

\item{ylab}{option to change y-axis label}

\item{xlab}{option to change x-axis label}

\item{plot.cex}{graphic option}

\item{index}{names of indices to plot (default = "all")}

\item{index.label}{show index name in plot}

\item{add}{if TRUE par is not called to enable manual multiplots}

\item{verbose}{commentary}
}
\value{
Bayesin p values and n observation per index
}
\description{
Plots Posterior Predictive Checks for Chisq Goodness-of-Fit (Omnibus Discrepancy)
}
