% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbplot_ppderived.R
\name{jbplot_ppderived}
\alias{jbplot_ppderived}
\title{Prior and posterior distributions of derived quantities: MSY, BMSY, HMSY, and BMSY/K}
\usage{
jbplot_ppderived(
  jabba,
  quants = c("MSY", "Bmsy", "Fmsy", "BmsyK"),
  output.dir = getwd(),
  as.png = FALSE,
  mfrow = c(round((ncol(jabba$refpts_posterior))/3 + 0.33, 0), 3),
  width = 8,
  height = 2.5 * round(ncol(jabba$refpts_posterior)/3, 0),
  cex = 0.8,
  verbose = TRUE,
  addPP = TRUE
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{quants}{vector of derived quantities to plot ("MSY", "BMsy", "Fmsy", "BmsyK")}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{mfrow}{set up plot frame}

\item{width}{plot width}

\item{height}{plot hight}

\item{verbose}{silent option}

\item{addPP}{show PPMR and PPVR}

\item{add}{if true don't call par() to allow construction of multiplots}
}
\description{
Prior and posterior distributions of derived quantities: MSY, BMSY, HMSY, and BMSY/K
}
