% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_ppdist}
\alias{jbplot_ppdist}
\title{Plot of prior and posterior distributions}
\usage{
jbplot_ppdist(
  jabba,
  output.dir = getwd(),
  as.png = FALSE,
  mfrow = c(round((ncol(jabba$pars_posterior))/3 + 0.33, 0), 3),
  width = 8,
  height = 2.5 * round(ncol(jabba$pars_posterior)/3, 0)
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{mfrow}{set up plot frame}

\item{width}{plot width}

\item{height}{plot hight}

\item{add}{if true don't call par() to allow construction of multiplots}
}
\description{
Prior and posterior distributions of esimated parameters: K, r, psi (depletion) and variances
}
