% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_prj}
\alias{jbplot_prj}
\title{Projection plot}
\usage{
jbplot_prj(
  jabba,
  type = c("BB0", "BBmsy", "FFmsy"),
  CIs = TRUE,
  flim = 6,
  output.dir = getwd(),
  as.png = FALSE,
  add = FALSE,
  mfrow = c(1, 1),
  width = 5,
  height = 3.5
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{type}{Options are c("BBmsy","FFmsy","BB0")}

\item{CIs}{Option to show CIs}

\item{flim}{max ylim value for FFmsy plot}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{add}{if true don't call par() to allow construction of multiplots}

\item{mfrow}{set up plot frame}

\item{width}{plot width}

\item{height}{plot hight}
}
\description{
plots choice of projections of B/Bmsy, F/Fmsy or B/B0 over fixed quotas set up in build_jabba()
}
