% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_procdev}
\alias{jbplot_procdev}
\title{Plot of process error deviation on log(biomass)}
\usage{
jbplot_procdev(
  jabba,
  output.dir = getwd(),
  as.png = FALSE,
  add = FALSE,
  width = 5,
  height = 3.5
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{add}{if true don't call par() to allow construction of multiplots}

\item{width}{plot width}

\item{height}{plot hight}
}
\description{
shows the difference of the expected biomass and its stochastic realization
}
