% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_retro}
\alias{jbplot_retro}
\title{jbplot_retro() to plot retrospective pattern}
\usage{
jbplot_retro(
  hc,
  type = c("B", "F", "BBmsy", "FFmsy", "procB", "SP"),
  forecast = FALSE,
  add = F,
  output.dir = getwd(),
  as.png = FALSE,
  single.plots = FALSE,
  width = NULL,
  height = NULL,
  xlim = NULL,
  cols = NULL,
  legend.loc = "topright",
  verbose = TRUE
)
}
\arguments{
\item{hc}{output list from hindast_jabba()}

\item{type}{single plot option c("B","F","BBmsy","FFmsy","BB0","SP")}

\item{forecast}{includes retrospective forecasting if TRUE}

\item{add}{add to multi plot if TRUE}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{width}{plot width}

\item{height}{plot hight}

\item{xlim}{allows to "zoom-in" requires speficiation Xlim=c(first.yr,last.yr)}

\item{cols}{option to add colour palette}

\item{legend.loc}{location of legend}

\item{verbose}{if FALSE be silent}
}
\value{
Mohn's rho statistic for several quantaties
}
\description{
Plots retrospective pattern of B, F, BBmsy, FFmsy, BB0 and SP #'
}
\examples{
data(iccat)
bet= iccat$bet
jb = build_jabba(catch=bet$catch,cpue=bet$cpue,se=bet$se,assessment="BET",scenario = "Ref",model.type = "Pella",igamma = c(0.001,0.001),verbose=FALSE)
fit = fit_jabba(jb,quickmcmc=TRUE,verbose=FALSE)
hc = hindcast_jabba(jbinput=jb,fit=fit,peels=1:3)
jbplot_retro(hc)
jbplot_retro(hc,forecast=TRUE) # with retro forecasting
}
