% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_spdyn}
\alias{jbplot_spdyn}
\title{JABBA SPdyn Plot}
\usage{
jbplot_spdyn(
  jabba,
  output.dir = getwd(),
  as.png = FALSE,
  add = FALSE,
  width = 5,
  height = 4.5
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{add}{if true don't call par() to allow construction of multiplots}

\item{width}{plot width}

\item{height}{plot hight}
}
\description{
plots the production vs biomass (Walters et al 2008) and color-coded kobe phases
}
