% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_spphase}
\alias{jbplot_spphase}
\title{JABBA SPphase Plot}
\usage{
jbplot_spphase(
  jabba,
  output.dir = getwd(),
  as.png = FALSE,
  add = FALSE,
  width = 5,
  height = 4.5,
  verbose = TRUE
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{add}{if true don't call par() to allow construction of multiplots}

\item{width}{plot width}

\item{height}{plot hight}

\item{verbose}{silent option}
}
\description{
plots the production function + Catch vs biomass and color-coded kobe phases
}
