% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_TOE}
\alias{jbplot_TOE}
\title{Stacked total observation plot}
\usage{
jbplot_TOE(
  jabba,
  ylab = NULL,
  xlab = NULL,
  output.dir = getwd(),
  as.png = FALSE,
  add = FALSE,
  width = 5,
  height = 4.5,
  verbose = TRUE
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{ylab}{yaxis label, default "Observation Error Variance"}

\item{xlab}{xaxis label, default "Year"}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{add}{if true don't call par() to allow construction of multiplots}

\item{width}{plot width}

\item{height}{plot height}

\item{verbose}{silent option}
}
\description{
Stacked total observation plot
}
