% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_trj}
\alias{jbplot_trj}
\title{Plot of estimated trajectories}
\usage{
jbplot_trj(
  jabba,
  type = c("B", "F", "BBmsy", "FFmsy", "BB0"),
  output.dir = getwd(),
  as.png = FALSE,
  add = FALSE,
  mfrow = c(1, 1),
  width = 5,
  height = 3.5
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{type}{= c("B","F","BBmsy","FFmsy","BB0")}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{add}{if true don't call par() to allow construction of multiplots}

\item{mfrow}{set up plot frame}

\item{width}{plot width}

\item{height}{plot hight}
}
\description{
plots choice of trajectories of Biomass, F, B/Bmsy, F/Fmsy or B/B0
}
