% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_utils.R
\name{jbretro}
\alias{jbretro}
\title{jbretro() computes Mohn's rho and forecast rho}
\usage{
jbretro(
  hc,
  type = c("B", "F", "BBmsy", "FFmsy", "procB", "SP"),
  forecast = TRUE
)
}
\arguments{
\item{hc}{output list from hindast_jabba()}

\item{type}{option c("B","F","BBmsy","FFmsy","BB0","SP")}

\item{forecast}{includes retrospective forecasting if TRUE}
}
\value{
Mohn's rho statistic for several quantities
}
\description{
Quantities retrospective pattern of B, F, BBmsy, FFmsy, BB0 and SP #'
}
\examples{
data(iccat)
bet <- iccat$bet
jb <- build_jabba(catch=bet$catch,
                 cpue=bet$cpue,
                 se=bet$se,
                 assessment="BET",
                 scenario = "Ref",
                 model.type = "Pella",
                 igamma = c(0.001,0.001),
                 verbose=FALSE)
fit <- fit_jabba(jb,quickmcmc=TRUE,verbose=FALSE)
hc <- hindcast_jabba(jbinput=jb,fit=fit,peels=1:3)
jbretro(hc)
jbplot_retro(hc)
jbplot_retro(hc,forecast=TRUE) # with retro forecasting
}
