% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{jbspr0}
\alias{jbspr0}
\title{spr0()}
\usage{
jbspr0(bio = jbio(), byage = FALSE)
}
\arguments{
\item{byage}{if TRUE it return spr0_at_age}

\item{jbio}{object list from jbio()}
}
\value{
spr0

spr0 value(s)
}
\description{
Function to compute spr0
}
