% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selex.R
\name{jselex}
\alias{jselex}
\title{jselex()}
\usage{
jselex(grow, selpars)
}
\arguments{
\item{grow}{growth object from \code{jgrow()}}

\item{selpars}{vector with selectivity parameters:
\itemize{
  \item \code{SL50}: length at 50\% selection
  \item \code{SL95}: length at 95\% selection
  \item \code{SLdesc}: mean of descending limb
  \item \code{CVdesc}: CV of descending limb
  \item \code{Smin}: minimum selectivity on descending limb
}}
}
\value{
list with Sel and selpars
}
\description{
Computes selectivity-at-age from growth
}
\examples{
g <- jgrow()
  
sel <- jselex(g, selpars = c(35,45,70,0.15,0.05))
head(sel$Sel)
}
