% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{junction.paths}
\alias{junction.paths}
\title{junction.paths}
\usage{
junction.paths(cn, adj)
}
\arguments{
\item{cn}{length n vector of integer copy numbers}

\item{adj}{nxn matrix of junction copy numbers}
}
\value{
list
with fields
$paths  list of n paths, each path i consisting of an n_i x 2 matrix specifying sequences of n_i junctions (each an ij node pair)
$mcn minimum copy number associated with path i
}
\description{
Applies "pigeonhole principle" to enumerate all junction paths
in a karyograph that can be proven to have copy number greater than 0
}
\details{
Takes as input adjacency matrix specifying junction copy numbers
and numeric vector specifying node copy numbers.
}
