% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{karyograph}
\alias{karyograph}
\title{karyograph}
\usage{
karyograph(junctions, tile = NULL, label.edges = FALSE)
}
\arguments{
\item{junctions}{GRangesList of junctions, where each item is a length GRanges of signed locations}

\item{tile}{GRanges optional existing tiling of the genome (eg a copy number segmentation) from which additional segments will be created}
}
\value{
a list with the following fields
$tile = GRanges of length 2*n tiling of the genome corresponding to union of rearrangement breakpoints and copy number endpoints
$G = igraph object representing karyograph, here are the edge and vertex features 
     vertex features: $chrom, $start, $end, $width, $strand, $size, $shape, $border.width, $label, $chrom.ord, $y, $col, $weight
     edge features:$ $bp.id, $weight, $from, $to, $col, $type, $line.style, $arrow.shape, $width, 
     important:  $type specifies which edges are "aberrant" and "reference", $bp.id specifies which input rearrangement (item in junctions)
     a given aberrant edge came from (and is NA for reference edges)
$adj = 2n x 2n adjacency matrix whose nonzero entries ij show the edge.id in $G
$ab.adj = 2n x 2n binary matrix specifying aberrant edges
$ab.edges = length(junctions) x {'from', 'to'} x {'+', '-'} mapping junction id's (indices into input junctions lists) to source and sink vertices,
            in both orientations
}
\description{
karyograph
}
\details{
builds graph from rearrangement breakpoints +/- copy number endpoints
used for downstream jbaMIP and karyoMIP functions

Input bpp is a GRangesList of signed locus pairs describing aberrant adjacencies.
The convention is as follows: Each locus in the input breakpoint pair points to the direction that
is being joined by the adjacencies i.e.
(-) bp points to "left" or preceding segment 
(+) bp points to the  "right" or the following segment

eg imagine a|bp1|b
           c|bp2|d
 "+" bp point to the right (eg b or d), "-" bp point to the left (a or c)

Input "tile" is a set of intervals whose endpoints are also used to partition the genome prior to the building of the
karyograph.

Output karyograph connects signed genomic intervals (in a signed tiling of the reference genome) with "aberrant" and "reference" edges.
Reference edges connect intervals that are adjacent in the reference genome, and aberrant edges are inferred (upstream
of this) through cancer genome paired end analysis.
Note that every node, edge, and path in this karyograph has a "reciprocal path"
}
