% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{karyograph_stub}
\alias{karyograph_stub}
\title{karyograph}
\usage{
karyograph_stub(seg.file, cov.file, nseg.file = NULL, het.file = NULL,
  ra = NULL, dranger.file = NULL, out.file, ra.file = NULL,
  force.seqlengths = NULL, purity = NA, ploidy = NA, field = "ratio",
  mc.cores = 1, max.chunk = 1e+08, subsample = NULL)
}
\description{
karyograph
}
