% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{karyoMIP}
\alias{karyoMIP}
\title{karyoMIP}
\usage{
karyoMIP(K, e, eclass = 1:length(e), kclass = NULL, prior = rep(0,
  ncol(K)), cpenalty = 1, tilim = 100, epgap = 1, nsolutions = 50,
  objsense = "max", ...)
}
\arguments{
\item{K}{|E| x k binary matrix of k "extreme" contigs across |E| edges}

\item{e}{edge copy numbers across |E| edges}

\item{eclass}{edge equivalence classes, used to constrain strand flipped contigs to appear in solutions together, each class can have at most 2 members}

\item{prior}{prior log likelihood of a given contig being in the karyotype}

\item{cpenalty}{karyotype complexity penalty - log likelihood penalty given to having a novel contig in the karyotype, should be calibrated to prior, ie higher than the contig-contig variance in the prior, otherwise complex karyotypes may be favored}

\item{tilim}{time limit to optimizatoin}

\item{nsolutions}{how many equivalent solutions to report}
}
\value{
Rcplex solution list object with additional field $kcn for path copy number, $kclass for k class id, $mval for mval
}
\description{
MIP to locally compute walks in an existing JaBbA reconstruction, note: usually many optimal solutions to a given run.
Used by jabba.walk.
}
\details{
TODO: Make user friendly, still pretty raw

takes |E| x k matrix of k extreme paths (ie contigs) across e edges of the karyograph
and length |E| vector of edge copy numbers (eclass), length |E| vector of edge equivalence classes (both outputs of jbaMIP.process)
and computes most likely karyotypes that fit the edge copy number profile subject to some prior likelihood
over the k extreme paths
}

