% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{karyoMIP.to.path}
\alias{karyoMIP.to.path}
\title{karyoMIP.to.path}
\usage{
karyoMIP.to.path(sol, K, e, gr = NULL, mc.cores = 1, verbose = T)
}
\arguments{
\item{sol}{solution to karyoMIP}

\item{K}{matrix of elementary paths (input to karyoMIP)}

\item{e}{nrow(K) x 2 edge matrix representing vertex pairs (i.e. edges to which K is referring to)}

\item{gr}{optional GRanges whose names are indexed by rownames of B}

\item{mc.cores}{integer number of cores}

\item{verbose}{flag}
}
\value{
A list with following items: 
$path length k list of paths, cycles (each item i is vector denoting sequence of vertices in G )
$is.cycle length k logical vector whose component i denotes whether path i is cyclic
$cn  length k integer vector whose component i denotes copy number of contig i
$path.grl if path.grl == T
}
\description{
for a karyoMIP solution and associated K matrix of n x e elementary paths  (input to karyoMIP), and v x e edge signed incidence matrix
}

