% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{karyoseg}
\alias{karyoseg}
\title{karyoseg}
\usage{
karyoseg(kag, cov)
}
\description{
performs preliminary segmentation on preliminary karyograph using CBS on coverage in tiles of coverage samples on the genome
only segments inter-junction reference intervals above min.width (which is usually defined as 20 times the coverage width)
returns the new karyograph with the addition partitions (and reference interval pairs)
}
