% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{karyotrack}
\alias{karyotrack}
\title{karyo track}
\usage{
karyotrack(kag, paths = NULL, col = "red", pad = 0)
}
\arguments{
\item{kag}{output of karyograph}

\item{paths}{GRanges or GRangesList}
}
\value{
gTrack of karyograph with particular nodes / edges colored with specified colors
}
\details{
Takes karyograph and outputs gTrack +/- highlighting of one or more paths defined as GRanges or GRangesList (for multiple paths)
Edges will only be highlighted when the exact interval pair corresponding to the edge is included in
the graph
}

