% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_utils.R
\name{kobeJabba}
\alias{kobeJabba}
\title{Function kobeJabba for FLR}
\usage{
kobeJabba(x)
}
\arguments{
\item{x}{posterior array dims(iter,year,stock,harvest)}
}
\description{
Function to convert kobe posteriors into KOBE FLR input object
}
