% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_utils.R
\name{kobeJabbaProj}
\alias{kobeJabbaProj}
\title{Function to convert JABBA projections into  FLR object}
\usage{
kobeJabbaProj(x)
}
\arguments{
\item{x}{posterior array dims(iter,year,tac,stock,harvest)}
}
\description{
Function to convert kobe projection matrix posteriors into Kobe FLR input object
}
