% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{loose.ends}
\alias{loose.ends}
\title{loose.ends}
\usage{
loose.ends(sol, kag)
}
\arguments{
\item{sol}{JaBbA object}

\item{kag}{karyograph object}
}
\value{
vector of ranges
on the right or left end of the intervals
that have type 1-4 labels where
}
\description{
takes jbaMIP output and outputs a vector of ranges
on the right or left end of the intervals
that have type 1-4 labels where
}
\details{
type1 = cn drop, no junction on this side, slack
type2 = cn drop, no junction on this side slack
type3 = no cn diff, used junction on other side, slack
type4 = no cn diff, unused junction on other side, no slack
}
