% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_jabba.R
\name{mp_jabba}
\alias{mp_jabba}
\title{Fit JABBA model}
\usage{
mp_jabba(
  jbinput,
  ni = 30000,
  nt = 5,
  nb = 5000,
  nc = 2,
  init.values = FALSE,
  init.K = NULL,
  init.r = NULL,
  init.q = NULL,
  par.quantile = 0.5,
  b.quantile = 0.5,
  quickmcmc = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{jbinput}{List of input variables as output by build_jabba()
MCMC settings}

\item{ni}{number of iterations}

\item{nt}{thinning interval of saved iterations}

\item{nb}{burn-in}

\item{nc}{number of mcmc chains
Initial values}

\item{init.values}{= FALSE,}

\item{init.K}{= NULL,}

\item{init.r}{= NULL,}

\item{init.q}{= NULL,# vector}

\item{par.quantile}{quantile(s) of parameter posterior, default median 0.5}

\item{b.quantile}{quantile(s) of biomass posterior, default median 0.5}

\item{quickmcmc}{option to run short mcmc}

\item{verbose}{option show cat comments and progress}
}
\value{
A result list containing estimates of model input, settings and results
}
\description{
Fits JABBA model in JAGS and produce output object as list()
}
\examples{
data(iccat)
jbinput <- build_jabba(catch=iccat$bet$catch,
                       cpue=iccat$bet$cpue,
                       se=iccat$bet$se,
                       model.type="Fox",
                       verbose=FALSE)
system.time(fit_jabba(jbinput,quickmcmc=TRUE,verbose=FALSE))
system.time(mp_jabba(jbinput))
}
