% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_utils.R
\name{normIndex}
\alias{normIndex}
\title{normIndex()}
\usage{
normIndex(cpue)
}
\arguments{
\item{cpue}{first column must be year}
}
\value{
normalized cpue
}
\description{
Function function to normalize CPUE indices
}
