% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{om}
\alias{om}
\title{om}
\format{
List of 6
\describe{
\item{stock}{data.frame of 130 rows and 7 cols}
\item{harvest}{data.frame of 130 rows and 7 cols}
\item{catch}{data.frame of 130 rows and 7 cols}
\item{fmsy}{fmsy}
\item{bmsy}{bmsy}
\item{msy}{msy}
}
}
\usage{
om
}
\description{
poly.sim om
}
\keyword{datasets}
