% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_utils.R
\name{plot_lnorm}
\alias{plot_lnorm}
\title{gets lognormal prior paramenters}
\usage{
plot_lnorm(mu, CV, Prior = "x", Plot = FALSE)

plot_lnorm(mu, CV, Prior = "x", Plot = FALSE)
}
\arguments{
\item{mu}{mean of lognormal distribution}

\item{CV}{cv of lognoram distribution}

\item{Prior}{X-axis label}

\item{Plot}{c(TRUE,FALSE)}
}
\value{
mean and lod.sd

mean and lod.sd
}
\description{
gets lognormal prior paramenters mean and log.sd with bias correction and plots distribution

Gets lognormal prior paramenters mean and log.sd with bias correction and plots distribution
}
