% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{pp.nll}
\alias{pp.nll}
\title{pp.nll}
\usage{
pp.nll(segstats, purity = NA, ploidy = NA, gamma = NA, beta = NA,
  field = "mean", field.ncn = "ncn", v = NULL)
}
\arguments{
\item{segstats}{granges of segstats with field $mean and $sd corresponding to mean and standard deviation on estimates of fragment density}

\item{purity}{numeric purity value}

\item{ploidy}{numeric ploidy value}

\item{gamma}{numeric gamma value (over-rides purity, ploidy)}

\item{beta}{numeric beta value (over-rides purity, ploidy)}

\item{field.ncn}{character specifying meta data field specifying germline copy number, default value is "ncn", if absent then will assume 2}

\item{solution}{over-ride if not interested in MLE / weighted least squares fit, this should be an integer vector of length(segstats)}
}
\value{
negative log likelihood of MLE i.e. least squares model (or supplied solution) fit
}
\description{
computes neg log likelihood ($nll) for purity, ploidy combo and mle abs copy numbers ($v), returns as list
}
\details{
v can be over-ridden to compute NLL for other (eg non MLE) values
}

