% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{pp2gb}
\alias{pp2gb}
\title{pp2gb}
\usage{
pp2gb(purity = NA, ploidy = NA, mu = NA, w = NA, gamma = NA,
  beta = NA)
}
\arguments{
\item{purity}{value between 0 and 1}

\item{ploidy}{value nonnegative}

\item{mu}{vector of n segment averages}

\item{w}{vector of n segment widths}

\item{gamma}{non-negative value}

\item{beta}{non-negative value}
}
\value{
list with purity / ploidy / gamma / beta entries
}
\description{
converts purity / ploidy to gamma / beta (or reverse)
}
\details{
takes in gr with signal field "field"
}

