% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{ppgrid}
\alias{ppgrid}
\title{ppgrid}
\usage{
ppgrid(segstats, allelic = FALSE, purity.min = 0.01, purity.max = 1,
  ploidy.step = 0.01, purity.step = 0.01, ploidy.min = 1.2,
  ploidy.max = 6, plot = F, verbose = F, mc.cores = 10)
}
\arguments{
\item{segstats}{GRanges object of intervals with meta data fields "mean" and "sd" (i.e. output of segstats function)}

\item{allelic}{logical flag, if TRUE will also look for mean_high, sd_high, mean_low, sd_low variables and choose among top solutions from top copy number according to the best allelic fit}

\item{mc.cores}{integer number of cores to use (default 1)}
}
\value{
data.frame with top purity and ploidy solutions and associated gamma and beta values, for use in downstream jbaMIP
}
\description{
least squares grid search for purity and ploidy modes
}

