% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{proximity}
\alias{proximity}
\title{proximity}
\usage{
proximity(query, subject, ra = GRangesList(), jab = NULL, verbose = F,
  mc.cores = 1, max.dist = 1e+06)
}
\arguments{
\item{query}{GRanges of "intervals of interest" eg regulatory elements}

\item{subject}{GRanges of "intervals of interest" eg genes}

\item{ra}{GRangesList of junctions (each a length 2 GRanges, similar to input to karyograph)}

\item{jab}{existing JaBbA object (overrides ra input)}

\item{verbose}{logical flag}

\item{mc.cores}{how many cores (default 1)}

\item{max.dist}{maximum genomic distance to store and compute (1MB by default) should the maximum distance at which biological interactions may occur}
}
\value{
list of n x m sparse distance matrices:
$ra = subject-query distance in the rearranged genome for all loci < max.dist in tumor genome
$wt = subject-query distance in the reference genome for all loci < max.dist in tumor genome
$rel = subject-query distance in ra relative to wild type for above loci
NOTE: values x_ij in these matrices should be interpreted with a 1e-9 offset to yield the actual value y_ij
i.e. y_ij = x_ij-1e-9, x_ij>0, y_ij = NA otherwise (allows for sparse encoding of giant matrices)
}
\description{
Takes a set of n "query" elements (GRanges object, e.g. genes) and determines their proximity to m "subject" elements
(GRanges object, e.g. regulatory elements) subject to set of rearrangement adjacencies (GRangesList with width 1 range pairs)
}
\details{
This analysis makes the (pretty liberal) assumption that all pairs of adjacencies that can be linked on a karyograph path are in
cis (i.e. share a chromosome) in the tumor genome.
}

