% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsy2jabba.R
\name{range2prior}
\alias{range2prior}
\title{range2prior()}
\usage{
range2prior(lo, hi)
}
\arguments{
\item{lo}{lower limit}

\item{hi}{upper limit}
}
\value{
mu,cv and logsd of range limits
}
\description{
approximates mu, cv, logsd from a range limits
}
