% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_utils.R
\name{rc4}
\alias{rc4}
\title{rc4 r4ss color palette}
\usage{
rc4(n, alpha = 1)
}
\arguments{
\item{n}{number of colors}

\item{alpha}{transluscency}
}
\value{
vector of color codes
}
\description{
rc4 r4ss color palette
}
