% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{Rcplex2}
\alias{Rcplex2}
\title{over-ride Rcplex}
\usage{
Rcplex2(cvec, Amat, bvec, Qmat = NULL, lb = 0, ub = Inf,
  control = list(), objsense = c("min", "max"), sense = "L",
  vtype = NULL, n = 1)
}
\description{
over-ride Rcplex
}
