% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{reads.to.walks}
\alias{reads.to.walks}
\title{reads.to.walks}
\usage{
reads.to.walks(bam, walks, outdir = "./test", hg = skidb::read_hg(fft = T),
  mc.cores = 1, insert.size = 1000, verbose = T)
}
\arguments{
\item{bam}{path to bam file}

\item{walks}{GRangesList of walks}

\item{outdir}{outdir to compute into}

\item{hg}{human genome sequence BSGenome or ffTrack}

\item{mc.cores}{number of cores}

\item{insert.size}{>= max insert size of library so that all relevant read pairs are extracted from the original bam}
}
\value{
paths to new bams
These bams are in "walk coordinates"
}
\description{
Utility function to realign reads to walks.
}
\details{
Takes bam and collection of walks (GRanges list of signed intervals on hg19 or other BSgenome)
pulls reads in regions of walks, then uses bwa mem to realign reads to walks, returns paths to new bams
These bams are in "walk coordinates"

Assumes bwa mem installed an runnable from command line.
}
