% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{rel2abs}
\alias{rel2abs}
\title{rel2abs}
\usage{
rel2abs(gr, purity = NA, ploidy = NA, gamma = NA, beta = NA,
  field = "ratio", field.ncn = "ncn")
}
\arguments{
\item{gr}{GRanges input with meta data field corresponding to mean relative copy "mean" in that interval}

\item{purity}{purity of sample}

\item{ploidy}{ploidy of sample}

\item{gamma}{gamma fit of solution (over-rides purity and ploidy)}

\item{beta}{beta fit of solution (over-rides purity and ploidy)}

\item{field}{meta data field in "gr" variable from which to extract signal, default "mean"}

\item{field.ncn}{meta data field in "gr" variable from which to extract germline integer copy number, default "ncn", if doesn't exist, germline copy number is assumed to be zero}
}
\value{
numeric vector of integer copy numbers
}
\description{
rescales CN values from relative to "absolute" (i.e. per cancer cell copy) scale given purity and ploidy
}
\details{
takes in gr with signal field "field"
}
