% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_utils.R
\name{runs_sig3}
\alias{runs_sig3}
\title{Function to do runs.test and 3 x sigma limits}
\usage{
runs_sig3(x, type = NULL)
}
\arguments{
\item{x}{residuals from CPUE fits}

\item{type}{only c("resid","observations")}
}
\value{
runs p value and 3 x sigma limits
}
\description{
runs test is conducted with library(snpar)
}
