% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{seg_stub}
\alias{seg_stub}
\title{segment
   this.cov = seg2gr(read.delim(cov.file, strings = F), seqlengths = hg_seqlengths())}
\usage{
seg_stub(cov.file, out.file, field = "ratio", mc.cores = 4)
}
\description{
segment
   this.cov = seg2gr(read.delim(cov.file, strings = F), seqlengths = hg_seqlengths())
}

