% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{segstats}
\alias{segstats}
\title{segstats is a step in the JaBbA pipeline}
\usage{
segstats(target, signal = NULL, field = "signal", asignal = NULL,
  afields = c("ref.count", "alt.count"), prior_weight = 1,
  prior_mean = NA, prior_alpha = NA, prior_beta = NA, max.chunk = 1e+08,
  max.slice = 20000, na.thresh = 0.2, subsample = NULL, mc.cores = 1,
  nsamp_prior = 1000, ksamp_prior = 100)
}
\arguments{
\item{target}{GRanges of segments on which segstats will be computed}

\item{signal}{GRanges of coverage from which samples will be taken}

\item{field}{field of "signal" GRanges from which coverage signal will be pulled}

\item{asignal}{optional GRanges corresponding width 1 bialellic snp allele counts across the genome,}

\item{afields}{length 2 character vector meta data fields of asignal GRanges that will be used to get allele counts (default is ref.count, alt.count)}

\item{subsample}{number between 0 and 1 with which to subsample per segment for coverage (useful for superdense coverage eg 50 bases to avoid correlations between samples due to read overlap)}

\item{mc.cores}{number of cores to run on (default 1)}
}
\description{
segstats is a step in the JaBbA pipeline
}
\details{
computes posterior mean's and sd's for a target tiling GRanges of segments (target)
target must be a non-overlapping gapless strandless or two stranded tiling of the genome (eg output of gr.tile)
if two stranded, then every stranded interval must have a mirror image interval included 

given a GRanges of signals using value field "field" of signal GRanges
assuming that the signals inside each interval in "target" are independent samples from a
normal distribution of unknown mean and variance

will also compute means ands std deviations on a gamma posterior of the
the "high" and "low" alleles for a granges "asignal" representing allelic signal ref count and tot count
across a set of locations (i.e. modeling the counts as a poisson random variable)
Fields are specified by afields

outputs target GRanges with "$mean" and "$sd" fields populated
}

