% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{slack.breaks}
\alias{slack.breaks}
\title{slack.breaks}
\usage{
slack.breaks(segstats)
}
\description{
makes "breaks" GRL from slack edges in jbaMIP segstats output (GRanges object)
}
\details{
makes fake slack chromosomes corresponding to each nonzero slack edge
and connects points on the genome in the following orientation:

+ segment source slack --> - granges at endpoint of segment
+ segment target slack --> + granges at start point of segment
- segment source slack --> + granges at start point of segment
- segment target slack --> - granges at endpoint of segment
}
