% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_utils.R
\name{ss3col}
\alias{ss3col}
\title{ss3col r4ss color generator}
\usage{
ss3col(n, alpha = 1)
}
\arguments{
\item{n}{number of colors}

\item{alpha}{transluscency}
}
\value{
vector of color codes
}
\description{
ss3col r4ss color generator
}
