% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{superenhancer_stub}
\alias{superenhancer_stub}
\title{run super-enhancer analysis}
\usage{
superenhancer_stub(rg.span.file, se.bed.file, this.ra.file, out.file,
  verbose = T, max.dist = 1e+06)
}
\description{
run super-enhancer analysis
}

