% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_jabba.R
\name{fit_jabba}
\alias{fit_jabba}
\title{Fit JABBA model}
\usage{
fit_jabba(
  jbinput,
  ni = 30000,
  nt = 5,
  nb = 5000,
  nc = 2,
  init.values = FALSE,
  init.K = NULL,
  init.r = NULL,
  init.q = NULL,
  peels = NULL,
  do.ppc = TRUE,
  save.trj = TRUE,
  save.all = FALSE,
  save.jabba = FALSE,
  save.csvs = FALSE,
  output.dir = getwd(),
  quickmcmc = FALSE,
  seed = 123,
  jagsdir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{jbinput}{List of input variables as output by build_jabba()
MCMC settings}

\item{ni}{number of iterations}

\item{nt}{thinning interval of saved iterations}

\item{nb}{burn-in}

\item{nc}{number of mcmc chains
Initial values}

\item{init.values}{= FALSE,}

\item{init.K}{= NULL,}

\item{init.r}{= NULL,}

\item{init.q}{= NULL,# vector}

\item{peels}{= NULL, # retro peel option}

\item{do.ppc}{conducts and saves posterior predictive checks}

\item{save.trj}{adds posteriors of stock, harvest and bk trajectories}

\item{save.all}{add complete posteriors to fitted object}

\item{save.jabba}{saves jabba fit as rdata object}

\item{save.csvs}{option to write csv outputs}

\item{output.dir}{path to save plot. default is getwd()}

\item{quickmcmc}{option to run short mcmc}

\item{seed}{default 123, set random by e.g. sample.int(999,1)}

\item{verbose}{option show cat comments and progress}
}
\value{
A result list containing estimates of model input, settings and results
}
\description{
Fits JABBA model in JAGS and produce output object as list()
}
\examples{
\dontrun{
data(iccat)
jbinput <- build_jabba(catch=iccat$bet$catch,cpue=iccat$bet$cpue,se=iccat$bet$se,model.type="Fox")
bet1 = fit_jabba(jbinput,quickmcmc=TRUE,verbose=TRUE)
jbplot_summary(bet1)
}
}
