% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba2jags.R
\name{jabba2jags}
\alias{jabba2jags}
\title{Function to write JABBA model in JAGS}
\usage{
jabba2jags(jbinput)
}
\arguments{
\item{jbinput}{JABBA input object from build_jabba()}
}
\description{
Writes JAGS code for JABBA into a temporary directory
}
