% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jabba_plots}
\alias{jabba_plots}
\title{wrapper jbplot function}
\usage{
jabba_plots(
  jabba,
  output.dir = getwd(),
  as.png = TRUE,
  statusplot = "kobe",
  verbose = TRUE,
  bfrac = NULL
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{statusplot}{can be "kobe", or "bfrac". "kobe" produces the traditional quandrants, "bfrac" produces quandrants based on state-dependent reference points (added for PIFSC use)}

\item{verbose}{silent option}

\item{bfrac}{Fraction of BMSY to use for Kobe plot (e.g. bfrac = (1-M)), if including set statusplot to "bfrac"}
}
\description{
plots all routine JABBA plot to output.dir if as.png=TRUE (default)
}
