% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{jbio}
\alias{jbio}
\title{jbio()}
\usage{
jbio(
  amin = 0,
  amax = 10,
  nsexes = 2,
  nseas = 1,
  Loo = c(80, 65),
  k = c(0.2, 0.25),
  t0 = c(-0.5, -0.7),
  aW = 0.01,
  bW = 3.04,
  mat = c(40, 45, 0),
  M = c(0.3, 0.4),
  h = 0.7,
  fec = NA,
  empty = FALSE
)
}
\arguments{
\item{amin}{minimum age}

\item{amax}{maximum age}

\item{nsexes}{if nsexes = 2, values are in order female, male}

\item{Loo}{asymptotic length}

\item{k}{brody growth coefficient}

\item{t0}{theoretical age at zero length (vb growth)}

\item{aW}{parameter of length-weigth Wa = aW*La^bW}

\item{bW}{parameter of length-weigth Wa = aW*La^bW}

\item{mat}{c(lm,dm,0=len or 1 = age) logistic maturity ogive parameters
\itemize{
  \item m50 - age or length
  \item m95 - age or length
  \item P(mat) = 1/(1+exp(-log(19)*(x-m50)/(m95-x)))
  \item x - age or length
}}

\item{M}{Natural mortality (vectors of M at age can be adjusted after stk is created)}

\item{h}{steepness of BevHolt SSR}

\item{fec}{number of offspring per female - for sharks and rays}

\item{empty}{produces an empty list structure based on amin:amax if TRUE}
}
\value{
jbio List of life history vectors and matrices
}
\description{
Creates a list for JABBA-Select stock object as input for the ASEM
}
