% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_cpuefits}
\alias{jbplot_cpuefits}
\title{Plot of  fitted CPUE indices}
\usage{
jbplot_cpuefits(
  jabba,
  output.dir = getwd(),
  as.png = FALSE,
  single.plots = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{width}{plot width}

\item{height}{plot hight}
}
\description{
Plots observed and fitted cpue indices with expexted CIs (dark grey) and posterior predictive distribution (light grey)
}
