% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_kobe_bfrac}
\alias{jbplot_kobe_bfrac}
\title{Kobe plot with state-dependent reference point}
\usage{
jbplot_kobe_bfrac(
  jabba,
  bfrac = bfrac,
  ylab = NULL,
  xlab = NULL,
  output.dir = getwd(),
  as.png = FALSE,
  add = FALSE,
  width = 5,
  height = 4.5,
  verbose = TRUE
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{bfrac}{fraction of bmsy for reference point}

\item{ylab}{yaxis label}

\item{xlab}{xaxis label}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{add}{if true don't call par() to allow construction of multiplots}

\item{width}{plot width}

\item{height}{plot height}

\item{verbose}{silent option}
}
\description{
Used for PIFSC stocks where reference points are BMSY(1-M)
}
