% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_logfits}
\alias{jbplot_logfits}
\title{log(CPUE) fits}
\usage{
jbplot_logfits(
  jabba,
  output.dir = getwd(),
  as.png = FALSE,
  single.plots = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{jabba}{output list from fit_jabba}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{width}{plot width}

\item{height}{plot hight}
}
\description{
Plot of fitted CPUE indices on log-scale (r4ss-style)
}
