% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_retro}
\alias{jbplot_retro}
\title{jbplot_retro() to plot retrospective pattern}
\usage{
jbplot_retro(
  hc,
  output.dir = getwd(),
  as.png = FALSE,
  single.plots = FALSE,
  width = NULL,
  height = NULL,
  Xlim = NULL
)
}
\arguments{
\item{hc}{output from jabba_hindast()}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{width}{plot width}

\item{height}{plot hight}

\item{Xlim}{allows to "zoom-in" requires speficiation Xlim=c(first.yr,last.yr)}
}
\value{
Mohn's rho statistic for several quantaties
}
\description{
Plots retrospective pattern of B, F, BBmsy, FFmsy, BB0 and SP #'
}
