% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_summary}
\alias{jbplot_summary}
\title{jbplot_summary()}
\usage{
jbplot_summary(
  jabbas,
  type = c("B", "F", "BBmsy", "FFmsy", "BB0", "SP"),
  plotCIs = TRUE,
  ylabs = NULL,
  prefix = "Summary",
  save.summary = FALSE,
  output.dir = getwd(),
  as.png = FALSE,
  single.plots = add,
  add = FALSE,
  width = NULL,
  height = NULL,
  xlim = NULL,
  cols = NULL,
  legend.loc = "topright",
  legend.cex = 0.8,
  legend.add = TRUE,
  plot.cex = 0.8,
  verbose = TRUE
)
}
\arguments{
\item{jabbas}{list() of JABBA model 1:n}

\item{type}{for single plots optional select type=c("B","F","BBmsy","FFmsy","BB0","SP")}

\item{plotCIs}{Plot Credibilty Interval}

\item{ylabs}{yaxis labels for quants}

\item{prefix}{Plot name specifier}

\item{save.summary}{option to save a summary of all loaded model runs}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{add}{adds single plots to mfrow set up}

\item{width}{plot width}

\item{height}{plot hight}

\item{xlim}{allows to "zoom-in" requires speficiation xlim=c(first.yr,last.yr)}

\item{cols}{option to add colour palette}

\item{legend.loc}{location of legend}

\item{legend.cex}{size of legend}

\item{legend.add}{show legend}

\item{plot.cex}{cex setting in par()}

\item{verbose}{if FALSE be silent}
}
\description{
Compares B, F, BBmsy, FFmsy, BB0 and SP for various model scanarios that have to be saved as rdata
}
