% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba_plots.R
\name{jbplot_summary}
\alias{jbplot_summary}
\title{jbplot_summary()}
\usage{
jbplot_summary(
  scenarios = NULL,
  assessment = NULL,
  mod.path = getwd(),
  plotCIs = TRUE,
  prefix = "Summary",
  save.summary = FALSE,
  output.dir = getwd(),
  as.png = FALSE,
  single.plots = FALSE,
  width = NULL,
  height = NULL,
  Xlim = NULL
)
}
\arguments{
\item{scenarios}{Names of model scenarios to compare}

\item{assessment}{Name of assessment as specified in jbinput}

\item{mod.path}{Directory of saved model runs or vector}

\item{plotCIs}{Plot Credibilty Interval}

\item{prefix}{Plot name specifier}

\item{save.summary}{option to save a summary of all loaded model runs}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{width}{plot width}

\item{height}{plot hight}

\item{Xlim}{allows to "zoom-in" requires speficiation Xlim=c(first.yr,last.yr)}
}
\description{
Compares B, F, BBmsy, FFmsy, BB0 and SP for various model scanarios that have to be saved as rdata
}
