% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selex.R
\name{jgrow}
\alias{jgrow}
\title{jgrow()}
\usage{
jgrow(
  param = list(Linf = 80, k = 0.2, t0 = -0.5),
  age = 0:12,
  nsexes = NULL,
  model = "vb"
)
}
\arguments{
\item{param}{list with growth parameters:
\itemize{
  \item \code{Linf}: asymptotic length
  \item \code{k}: growth coefficient
  \item \code{t0}: theoretical age at zero length
  \item \code{L0}: length at age 0, optional alternative to \code{t0}
}}

\item{age}{vector of ages}

\item{nsexes}{number of sexes}

\item{model}{growth model c("vb","schnute")}
}
\value{
list with age, nsexes, La, model and param
}
\description{
Computes length-at-age from a growth model
}
\examples{
pars <- list(
 Linf   = c(80,65),
  k     = c(0.2,0.25),
  t0     = c(-0.5,-0.7)
)
g <- jgrow(pars, model = "vb")
head(g$La)

}
