% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{karyograph_stub}
\alias{karyograph_stub}
\title{karyograph
     if (is.character(tryCatch(png(paste(out.file, '.ppgrid.png', sep = ''), height = 500, width = 500), error = function(e) 'bla')))}
\usage{
karyograph_stub(seg.file, cov.file, nseg.file = NULL, het.file = NULL,
  ra = NULL, dranger.file = NULL, out.file, ra.file = NULL,
  force.seqlengths = NULL, purity = NA, ploidy = NA, field = "ratio",
  mc.cores = 1, max.chunk = 1e+08, subsample = NULL)
}
\description{
karyograph
     if (is.character(tryCatch(png(paste(out.file, '.ppgrid.png', sep = ''), height = 500, width = 500), error = function(e) 'bla')))
}

