% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaBbA.R
\name{ramip_stub}
\alias{ramip_stub}
\title{run ramip}
\usage{
ramip_stub(kag.file, out.file, mc.cores = 1, mem = 16, tilim = 1200,
  slack.prior = 0.001, gamma = NA, beta = NA, customparams = T,
  edge.nudge = 0, ab.force = NULL)
}
\description{
run ramip
}

